proyecto base front, instalacion de SweetAlert, Notyf, agregado cdn de bootstrap y fontawesome
