+++ front/jsconfig.json
@@ -6,3 +6,5 @@
   },
   "exclude": ["node_modules", "dist"]
 }
+
+
