+++ front/package.json
@@ -10,7 +10,8 @@
   },
   "dependencies": {
     "axios": "^1.7.7",
-    "sweetalert2": "^11.14.2",
+    "notyf": "^3.10.0",
+    "sweetalert2": "^11.14.3",
     "vue": "^3.5.11",
     "vue-router": "^4.4.5"
   },
