+++ front/src/App.vue
@@ -1,0 +1,13 @@
+<script setup>
+import { RouterLink, RouterView } from 'vue-router'
+</script>
+
+<template>
+   
+
+  <RouterView />
+</template>
+
+<style scoped>
+
+</style>
