+++ front/src/assets/main.css
@@ -1,0 +1,2 @@
+@import './base.css';
+
