+++ front/src/main.js
@@ -1,0 +1,17 @@
+import './assets/main.css'
+
+import { createApp } from 'vue'
+import App from './App.vue'
+import router from './router'
+
+import Swal from 'sweetalert2';
+import { Notyf } from 'notyf';
+import 'notyf/notyf.min.css';
+
+const app = createApp(App)
+
+app.config.globalProperties.$swal = Swal;
+app.config.globalProperties.$notyf = new Notyf();
+
+app.use(router)
+app.mount('#app')
