+++ front/src/views/HomeView.vue
@@ -1,0 +1,8 @@
+<script setup>
+</script>
+
+<template>
+  <main >
+   Home
+  </main>
+</template>
